Extract work function helpers into shared library
